виртуальные функции + перегрузки

+++ triangle.cpp
@@ -19,7 +19,7 @@ void triangle::workTriangle()
     double p;
     perimeter = sides[0] + sides[1] + sides[2];
     p = perimeter / 2;
-    area = sqrtf(p * (p - sides[0]) * (p - sides[1]) * (p - sides[2]));
+    area = sqrt(p * (p - sides[0]) * (p - sides[1]) * (p - sides[2]));
     if (area == 0)
     {
         throw "Такого треугольника не существует";
